Split KotlinCoreEnvironmentWrapper file
